// sdk/src/dapp-api/rpc-gen/typings.ts
export type Id = string
/**
 *
 * The type of client that implements the wallet kernel.
 *
 */
export type ClientType = 'browser' | 'desktop' | 'mobile' | 'remote'
/**
 *
 * The URL of the wallet kernel.
 *
 */
export type Url = string
/**
 *
 * Represents a wallet kernel.
 *
 */
export interface KernelInfo {
    id: Id
    clientType: ClientType
    url?: Url
    [k: string]: any
}
/**
 *
 * Whether or not a connection to a network is esablished.
 *
 */
export type IsConnected = boolean
/**
 *
 * The network ID the wallet corresponds to.
 *
 */
export type ChainId = string
/**
 *
 * A URL that points to a user interface.
 *
 */
export type UserUrl = string
// ...
export type Response = string
/**
 *
 * JWT authentication token (if applicable).
 *
 */
export type SessionToken = string
/**
 *
 * Set as primary wallet for dApp usage.
 *
 */
export type Primary = boolean
/**
 *
 * The party ID corresponding to the wallet.
 *
 */
export type PartyId = string
/**
 *
 * The party hint and name of the wallet.
 *
 */
export type Hint = string
/**
 *
 * The public key of the party.
 *
 */
export type PublicKey = string
/**
 *
 * The namespace of the party.
 *
 */
export type Namespace = string
/**
 *
 * The signing provider ID the wallet corresponds to.
 *
 */
export type SigningProviderId = string
/**
 *
 * Structure representing a wallet
 *
 */
export interface Wallet {
    primary: Primary
    partyId: PartyId
    hint: Hint
    publicKey: PublicKey
    namespace: Namespace
    chainId: ChainId
    signingProviderId: SigningProviderId
    [k: string]: any
}
// ...
export interface StatusResult {
    kernel: KernelInfo
    isConnected: IsConnected
    chainId?: ChainId
    [k: string]: any
}
export interface ConnectResult {
    kernel: KernelInfo
    isConnected: IsConnected
    chainId?: ChainId
    userUrl: UserUrl
    [k: string]: any
}
export interface DarsAvailableResult {
    dars: Dars
    [k: string]: any
}
export type PrepareReturnResult = any
export interface PrepareExecuteResult {
    userUrl: UserUrl
    [k: string]: any
}
/**
 *
 * Ledger Api configuration options
 *
 */
export interface LedgerApiResult {
    response: Response
    [k: string]: any
}
export interface OnConnectedEvent {
    kernel: KernelInfo
    chainId: ChainId
    sessionToken?: SessionToken
    [k: string]: any
}
/**
 *
 * Event emitted when the user's accounts change.
 *
 */
export type AccountsChangedEvent = Wallet[]
/**
 *
 * An array of accounts that the user has authorized the dapp to access..
 *
 */
export type RequestAccountsResult = Wallet[]